etl clinical: refactor age funcs to shared module
Move the `age` and `age_to_delete` functions from the `etl clinical`
process to a shared module for resuse by future commands.
